+++ CHANGELOG.md
@@ -1,8 +1,18 @@
-## 0.0.1
+## Current
 
-* Initial release
+* "autolist" is a default setting now
+
+
+## 1.0.1
+
+* Corrected small bug preventing byebug from loading
 
 
 ## 1.0.0
 
 * Green test suite
+
+
+## 0.0.1
+
+* Initial release